requirements.txt
tensorflow-gpu==2.10.0
keras==2.10.0
scikit-learn==1.1.3
numpy==1.23.2
matplotlib==3.6.2
umap-learn==0.5.3
seaborn==0.12.1
pandas==1.5.2